backend/requirements.txt
asgiref==3.8.1
black @ git+https://github.com/psf/black@d0ff3bd6cb82e35b1529155d29fca2c13442e68d
click==8.1.8
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
PyJWT==2.9.0
pytokens==0.1.10
sqlparse==0.5.3
